refactor: move pulse service into pulse feature
