+++ backend/features/pulse/src/test/kotlin/com/moneat/services/PulseServiceTest.kt
@@ -16,13 +16,20 @@
 
 package com.moneat.services
 
+import com.moneat.shared.services.PulsePayload
 import com.moneat.shared.services.PulseService
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.cancel
 import kotlinx.coroutines.runBlocking
+import kotlinx.serialization.json.Json
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotNull
+import kotlin.test.assertNull
 import kotlin.test.assertTrue
+import kotlin.time.Duration.Companion.hours
 
 /**
  * PulseService collects telemetry for self-hosted deployments.
@@ -30,6 +37,52 @@ import kotlin.test.assertTrue
  * In typical test env SELF_HOSTED is false, so enabled=false and metrics=null.
  */
 class PulseServiceTest {
+
+    @Test
+    fun `PulsePayload serializes and deserializes correctly`() {
+        val payload = PulsePayload(
+            deploymentId = "test-id-123",
+            version = "v1.2.3",
+            cpuCount = 4,
+            memTotalBytes = 8_000_000_000,
+            memUsedBytes = 4_000_000_000,
+            osName = "Linux",
+            osArch = "amd64",
+            jvmVersion = "21",
+            projectCount = 10,
+            userCount = 5,
+            eventCount = 1000,
+            issueCount = 50,
+            selfHost = true,
+            sslEnabled = true
+        )
+
+        val json = Json { encodeDefaults = true }
+        val encoded = json.encodeToString(PulsePayload.serializer(), payload)
+        val decoded = json.decodeFromString(PulsePayload.serializer(), encoded)
+
+        assertEquals(payload.deploymentId, decoded.deploymentId)
+        assertEquals(payload.version, decoded.version)
+        assertEquals(payload.cpuCount, decoded.cpuCount)
+        assertEquals(payload.memTotalBytes, decoded.memTotalBytes)
+        assertEquals(payload.memUsedBytes, decoded.memUsedBytes)
+        assertEquals(payload.osName, decoded.osName)
+        assertEquals(payload.projectCount, decoded.projectCount)
+        assertEquals(payload.sslEnabled, decoded.sslEnabled)
+    }
+
+    @Test
+    fun `PulsePayload defaults are applied`() {
+        val payload = PulsePayload(deploymentId = "minimal")
+
+        assertEquals(0, payload.cpuCount)
+        assertEquals("", payload.version)
+        assertEquals(0L, payload.memTotalBytes)
+        assertEquals("", payload.osName)
+        assertEquals(0L, payload.projectCount)
+        assertTrue(payload.selfHost)
+        assertFalse(payload.sslEnabled)
+    }
 
     @Test
     fun `getStatus returns valid PulseStatus structure`(): Unit =
@@ -64,4 +117,27 @@ class PulseServiceTest {
             // When SELF_HOSTED=false (typical in tests), enabled=false
             assertEquals(PulseService.isEnabled(), status.enabled)
         }
+
+    @Test
+    fun `getStatus has null metrics when disabled`() =
+        runBlocking {
+            val status = PulseService.getStatus()
+
+            if (!status.enabled) {
+                assertNull(status.metrics)
+            }
+        }
+
+    @Test
+    fun `start and stop does not throw`() {
+        val service = PulseService(
+            interval = 1.hours,
+            endpoint = "http://localhost:1/noop"
+        )
+        val scope = CoroutineScope(Dispatchers.Default)
+
+        service.start(scope)
+        service.stop()
+        scope.cancel()
+    }
 }
